fix(json): transcode fix
